ThreadsScreen: Populate mutableStateListOf with dummy objects
We want to be able to access an object from a specific index based on
total number of pages which we get after first request.

However, we are fetching pages as we navigate the pages and inserting them
in the mutable list isn't possible after init because index can be empty.
Therefore, we add emptyList objects and replace them on querying pages.

Signed-off-by: Aayush Gupta <aayushgupta219@gmail.com>